+++ internal/database/sqliterepository.go
@@ -16,7 +16,12 @@ import (
 	"go.elastic.co/apm/v2"
 )
 
-const defaultMaxBulkAddBatch = 500
+const (
+	defaultMaxBulkAddBatch = 500
+
+	dataColumnName     = "data"
+	baseDataColumnName = "baseData"
+)
 
 var (
 	ErrDuplicate    = errors.New("record already exists")
@@ -40,8 +45,8 @@ var keys = []keyDefinition{
 	{"kibanaVersion", "TEXT NOT NULL"},
 	{"type", "TEXT NOT NULL"},
 	{"path", "TEXT NOT NULL"},
-	{"data", "BLOB NOT NULL"},
-	{"baseData", "BLOB NOT NULL"},
+	{dataColumnName, "BLOB NOT NULL"},
+	{baseDataColumnName, "BLOB NOT NULL"},
 }
 
 type SQLiteRepository struct {
@@ -271,19 +276,23 @@ func (r *SQLiteRepository) AllFunc(ctx context.Context, database string, whereOp
 	span.Context.SetLabel("database.path", r.File(ctx))
 	defer span.End()
 
-	useBaseData := whereOptions == nil || !whereOptions.UseFullData()
+	useJSONFields := !whereOptions.SkipJSONFields()
+	useBaseData := !whereOptions.UseFullData()
 
-	var getKeys []string
+	getKeys := make([]string, 0, len(keys))
 	var query strings.Builder
 	query.WriteString("SELECT ")
 	for _, k := range keys {
-		if k.Name == "data" && useBaseData {
+		switch {
+		case !useJSONFields && (k.Name == dataColumnName || k.Name == baseDataColumnName):
 			continue
-		}
-		if k.Name == "baseData" && !useBaseData {
+		case k.Name == dataColumnName && useBaseData:
 			continue
+		case k.Name == baseDataColumnName && !useBaseData:
+			continue
+		default:
+			getKeys = append(getKeys, k.Name)
 		}
-		getKeys = append(getKeys, k.Name)
 	}
 	query.WriteString(strings.Join(getKeys, ", "))
 	query.WriteString(" FROM ")
@@ -303,7 +312,7 @@ func (r *SQLiteRepository) AllFunc(ctx context.Context, database string, whereOp
 	// Reuse pkg variable since all fields are scanned into it
 	var pkg Package
 	for rows.Next() {
-		if err := rows.Scan(
+		columns := []any{
 			&pkg.Cursor,
 			&pkg.Name,
 			&pkg.Version,
@@ -313,9 +322,13 @@ func (r *SQLiteRepository) AllFunc(ctx context.Context, database string, whereOp
 			&pkg.KibanaVersion,
 			&pkg.Type,
 			&pkg.Path,
-			&pkg.Data, // this variable will be assigned to BaseData if useBaseData is true
-			// to avoid creting a new variable, we reuse pkg.Data
-		); err != nil {
+		}
+		if useJSONFields {
+			// this variable will be assigned to BaseData if useBaseData is true
+			// to avoid creating a new variable, we reuse pkg.Data
+			columns = append(columns, &pkg.Data)
+		}
+		if err := rows.Scan(columns...); err != nil {
 			return err
 		}
 		if useBaseData {
@@ -369,6 +382,7 @@ type SQLOptions struct {
 	CurrentCursor string
 
 	IncludeFullData bool // If true, the query will return the full data field instead of the base data field
+	SkipPackageData bool // If true, no need to retrieve Data nor BaseData fields
 }
 
 func (o *SQLOptions) Where() (string, []any) {
@@ -432,4 +446,11 @@ func (o *SQLOptions) UseFullData() bool {
 		return false
 	}
 	return o.IncludeFullData
+}
+
+func (o *SQLOptions) SkipJSONFields() bool {
+	if o == nil {
+		return false
+	}
+	return o.SkipPackageData
 }